chore[workers] Renamed utils to kebab case for standard
