+++ workers/tests/process-cron.test.ts
@@ -1,7 +1,7 @@
 import processCron from "../src/tasks/process-cron";
 import { JobHelpers, AddJobFunction, Job } from "graphile-worker";
 import { CronPayload, Event, FunctionPayload } from "../src/types/types";
-import log from "../src/utils/logUtils";
+import log from "../src/utils/log-utils";
 
 const validCronJob: CronPayload = { funcName: "my-function" };
